Add ability to handle empty variable values
Even though this is not quite something we would normally expect, there
is no good reason why we should prevent it. The change makes any
variables set with no tags set be interpreted as literals. As an
addition we also propagate errors in parsing better that were just
indicated by a non-helpful message before.