Update Jenkinsfile to have static stability levels
Since we want to hard-pin the current levels of coverage, Jenkinsfile
has been updated to reflect it. We also disable automatic health update
since that is known to cause issues with development branch builds (the
branch will fail to build if an earlier commit had a higher code
coverage result).