Add `portable` mode for provider search (#218)
* new: add `portable` mode for provider search

In order to allow `summon` to be used from arbitrary directories the
provider search path has been extended to also look in a `Providers`
sub-directory where the `summon` executalbe resides. With this `summon`
can be used more easily without installing it in a system path where
only administrators have write access.

Signed-off-by: Chris Drexler <ckolumbus@ac-drexler.de>

* doc:  add details about provider search algorithm

add details on how providers are searched and how
a migration from a system provider directory to a local
directory can be done.