CNJR-4257: Support homebrew provider dir
Currently the homebrew formula for summon-Conjur
installs within the brew directory to a location
that summon doesn't search (brew_root/lib/summon).

To support this we add the brew lib directory as
a search path.

This change also ensures that at least one
provider directory exists and fails early with a
helpful error message if a provider dir isn't
found. Previously the fallback providers dir was
assumed to exist but not checked.

Finally, when provider versions are listed, the
current provider path is shown to help users
understand where the versions are coming from.